feat: apply error intel + refactoring
